requirements.txt
opencv-python~=4.6.0.66
numpy~=1.23.4
torch~=1.12.1
torchvision~=0.13.1
requests~=2.28.1
pyserial~=3.5
geopy~=2.2.0